[#175848172] Moved narrow integration tests in __integrations__ folder
